+++ packages/stream_video_flutter/lib/src/call_screen/call_content/call_content.dart
@@ -146,6 +146,9 @@ class _StreamCallContentState extends State<StreamCallContent> {
                 configuration:
                     widget.pictureInPictureConfiguration.iOSPiPConfiguration,
                 participantSort: widget.pictureInPictureConfiguration.sort,
+                prioritiseScreenSharingTrack: widget
+                    .pictureInPictureConfiguration
+                    .prioritiseScreenSharingTrack,
               ),
             ),
           if (CurrentPlatform.isAndroid && pipEnabled)