+++ packages/stream_video_flutter/lib/src/call_screen/call_content/picture_in_picture/picture_in_picture_configuration.dart
@@ -14,6 +14,7 @@ class PictureInPictureConfiguration {
   const PictureInPictureConfiguration({
     this.enablePictureInPicture = false,
     this.disablePictureInPictureWhenScreenSharing = true,
+    this.prioritiseScreenSharingTrack = true,
     this.sort,
     this.androidPiPConfiguration = const AndroidPictureInPictureConfiguration(),
     this.iOSPiPConfiguration = const IOSPictureInPictureConfiguration(),
@@ -24,6 +25,16 @@ class PictureInPictureConfiguration {
 
   /// Whether to disable picture-in-picture mode during screen sharing on the device.
   final bool disablePictureInPictureWhenScreenSharing;
+
+  /// Whether to prioritise the screen sharing track over the camera track.
+  ///
+  /// If `true`, the screen sharing track will be displayed in the PiP view
+  /// if available for the first participant determined by the sorting function.
+  ///
+  /// If `false`, the camera track will be preferred. However, if the camera
+  /// track is disabled and screen sharing is available, the screen share
+  /// will be shown as a fallback.
+  final bool prioritiseScreenSharingTrack;
 
   /// Sorting function for participants in picture-in-picture mode.
   /// The first participant will be displayed in the PiP view.