+++ packages/stream_video_flutter/lib/src/call_screen/call_content/picture_in_picture/stream_picture_in_picture_android_view.dart
@@ -151,6 +151,8 @@ class _StreamPictureInPictureAndroidViewState
             .configuration
             .androidPiPConfiguration
             .callPictureInPictureWidgetBuilder,
+        prioritiseScreenSharingTrack:
+            widget.configuration.prioritiseScreenSharingTrack,
       ),
     );